Showing botColors inside the participants list
