+++ front-end/users/airo-users-frontend/src/app/home/components/events/components/event-live-feed/event-live-feed.component.html
@@ -13,7 +13,7 @@
                 <div class="participants">
                     <h3>Participants</h3>
                     @for (p of eventLiveFeed.simulationState.participants; track p.userId) {
-                    <p>{{p.userId}} - Health: {{p.health}}</p>
+                    <p [ngStyle]="{ 'backgroundColor': botColors[p.botId].toHexString() }">{{p.userId}} - Health: {{p.health}}</p>
                     }
                 </div>
                 <div class="log-container">
