+++ front-end/users/airo-users-frontend/src/app/home/components/events/components/event-live-feed/event-live-feed.component.ts
@@ -1,16 +1,19 @@
 import { Component, ViewChild } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { EventLiveFeedService, GetLiveFeedResponse, TileInfoDto } from '../../services/event-live-feed.service';
+import { EventLiveFeedService, GetLiveFeedResponse, Participant, TileInfoDto } from '../../services/event-live-feed.service';
 import { MapRendererComponent } from './map/map.component';
 import { LoadedMapData, TileInfo, TileType } from './map/models/map.models';
 import { CdkVirtualScrollViewport, ScrollingModule } from '@angular/cdk/scrolling';
+import { Color3 } from '@babylonjs/core';
+import { ColorDictionary } from './map/models/color-dictionary';
+import { NgStyle } from '@angular/common';
 
 @Component({
   selector: 'app-event-live-feed',
   templateUrl: './event-live-feed.component.html',
   styleUrl: './event-live-feed.component.scss',
   standalone: true,
-  imports: [MapRendererComponent, ScrollingModule]
+  imports: [MapRendererComponent, ScrollingModule, NgStyle]
 })
 export class EventLiveFeedComponent {
   @ViewChild('renderer', { static: false }) renderer!: MapRendererComponent;
@@ -21,7 +24,7 @@ export class EventLiveFeedComponent {
 
     try {
       if (this.eventLiveFeed) {
-        this.loadMap(this.eventLiveFeed.simulationState.tiles);
+        this.loadMap(this.eventLiveFeed.simulationState);
       }
 
     } catch (error) {
@@ -40,6 +43,15 @@ export class EventLiveFeedComponent {
   logs: string[] = [];
 
   autoScroll: boolean = true;
+
+  botColors: ColorDictionary = {};
+
+  private colors: Color3[] = [
+    Color3.Green(),
+    Color3.Red(),
+    Color3.Purple(),
+    Color3.Yellow(),
+  ];
 
   private lastReceived = 0;
   
@@ -71,7 +83,7 @@ export class EventLiveFeedComponent {
       this.eventLiveFeed = await this.getLiveFeed(this.lastReceived);
       if (this.eventLiveFeed) {
         this.updateLogs(this.eventLiveFeed.logs);
-        this.loadMap(this.eventLiveFeed.simulationState.tiles);
+        this.loadMap(this.eventLiveFeed.simulationState);
       }
     } catch (error) {
       console.error('Error fetching live feed:', error);
@@ -98,7 +110,7 @@ export class EventLiveFeedComponent {
     }
   }
 
-  private loadMap(tiles: TileInfoDto[][]) {
+  private loadMap({ participants, tiles }: { participants: Participant[]; tiles: TileInfoDto[][] }) {
     const mapData: LoadedMapData = {
       size: tiles.length,
       tiles: this.flattenTiles(tiles),
@@ -106,7 +118,11 @@ export class EventLiveFeedComponent {
 
     if (this.renderer) {
       if (!this.mapInitialized) {
-        this.renderer.initMap(mapData);
+        participants.forEach((participant, i) => {
+          this.botColors[participant.botId!] = this.colors[i];
+        });
+
+        this.renderer.initMap(mapData, this.botColors);
         this.mapInitialized = true;
       }
       this.renderer.updateMap(mapData);
