+++ front-end/users/airo-users-frontend/src/app/home/components/events/components/event-live-feed/map/models/color-dictionary.ts
@@ -1,0 +1,5 @@
+import { Color3 } from "@babylonjs/core";
+
+export interface ColorDictionary {
+    [key: string]: Color3;
+}
